re-ordering the structure for the template and context creation files
