+++ csss-site/src/elections/templates/elections/nominee_links/create_or_update_election/create_election_nominee_links.html
@@ -16,8 +16,8 @@
 				{%  include 'elections/webform_format/election_time.html' %}
 				{%  include 'elections/webform_format/election_type.html' %}
 				{%  include 'elections/webform_format/election_websurvey.html' %}
-				{%  include 'elections/nominee_links/election_nominee_names.html' %}<br>
-				{%  include 'elections/webform/submission_buttons.html' %}
+				{%  include 'elections/nominee_links/create_or_update_election/election_nominee_names.html' %}<br>
+				{%  include 'elections/webform_format/submission_buttons.html' %}
 			{%  endif %}
 		</form>
 	</div>
