+++ csss-site/src/elections/templates/elections/nominee_links/create_or_update_election/update_election/nominee_links_table.html
@@ -8,7 +8,7 @@
 		<th>Link For Nominee</th>
 		<th>Modify Nominee</th>
 	</tr>
-	{%  include 'elections/nominee_links/nominee_links_table/draft_nominee_links.html' %}
-	{%  include 'elections/nominee_links/nominee_links_table/final_nominee_links.html' %}
+	{%  include 'elections/nominee_links/create_or_update_election/update_election/nominee_links_table/draft_nominee_links.html' %}
+	{%  include 'elections/nominee_links/create_or_update_election/update_election/nominee_links_table/final_nominee_links.html' %}
 </table>
 {%  endif %}
