+++ csss-site/src/elections/templates/elections/webform/form__webform.html
@@ -10,11 +10,11 @@
 {%  endif %}
 {%  include 'elections/webform_format/election_type.html' %}
 {%  include 'elections/webform_format/election_websurvey.html' %}
-{%  include 'elections/webform/submission_buttons.html' %}
+{%  include 'elections/webform_format/submission_buttons.html' %}
 <br><br><br>
 <input type="button" value="Add another Nominee" onClick="addBlankNominee('{{nominee_div__name}}');">
 <div id="{{nominee_div__name}}"></div>
-{%  include 'elections/webform/submission_buttons.html' %}
+{%  include 'elections/webform_format/submission_buttons.html' %}
 <br><br><br>
 <input type="button" value="Add another Nominee" onClick="addBlankNominee('{{nominee_div__name}}');">
 <br>
