--- csss-site/src/elections/views/create_context/nominee_links/utils/display_errors_html.py
@@ -1,6 +1,0 @@
-from csss.views.views import ERROR_MESSAGES_KEY
-
-
-def create_context_for_display_errors_html(context, error_messages=None):
-    if error_messages is not None:
-        context[ERROR_MESSAGES_KEY] = error_messages
