+++ csss-site/src/elections/views/endpoints/nominee_links/create_election_nominee_links.py
@@ -5,7 +5,7 @@ from django.shortcuts import render
 
 from csss.views.context_creation.create_authenticated_contexts import create_context_for_election_officer
 from elections.views.Constants import TAB_STRING
-from elections.views.create_context.nominee_links.create_election_nominee_links_html import \
+from elections.views.create_context.nominee_links.create_or_update_election.create_election_nominee_links_html import \
     create_context_for_create_election_nominee_links_html
 from elections.views.create_election.nominee_links.process_new_election_and_nominee_links import \
     process_new_election_and_nominee_links
@@ -29,4 +29,7 @@ def display_and_process_html_for_new_nominee_links_election(request):
         return process_new_election_and_nominee_links(request, context)
     else:
         create_context_for_create_election_nominee_links_html(context, create_new_election=True)
-        return render(request, 'elections/create_election/create_election_nominee_links.html', context)
+        return render(
+            request, 'elections/nominee_links/create_or_update_election/create_election_nominee_links.html',
+            context
+        )
