+++ csss-site/src/elections/views/endpoints/nominee_links/display_and_process_html_for_nominee_links.py
@@ -29,7 +29,10 @@ def display_and_process_html_for_modification_of_election_and_nominee_links__nom
 
     if len(Election.objects.all().filter(slug=slug)) != 1:
         context[ERROR_MESSAGES_KEY] = [f"Received invalid Election slug of {slug}"]
-        return render(request, 'elections/update_election/update_election_nominee_links.html', context)
+        return render(
+            request, 'elections/nominee_links/create_or_update_election/update_election_nominee_links.html',
+            context
+        )
 
     process_election = (request.method == "POST")
     election = Election.objects.get(slug=slug)
