+++ csss-site/src/elections/views/endpoints/nominee_links/display_and_process_html_for_nominee_modification__nominee_link.py
@@ -36,8 +36,10 @@ def display_and_process_html_for_nominee_modification(request):
         error_message = [f"No election attached to Nominee Link {nominee_links[0]} detected in the request"]
     if error_message is not None:
         create_context_for_create_or_update_nominee__nominee_links_html(context, error_messages=[error_message])
-        return render(request,
-                      'elections/update_nominee/create_or_update_nominee__nominee_links.html', context)
+        return render(
+            request, 'elections/nominee_links/create_or_update_nominee/create_or_update_nominee__nominee_links.html',
+            context
+        )
 
     if (request.method == "POST") and (CREATE_OR_UPDATE_NOMINEE__NAME in request.POST):
         return process_nominee__nominee_links(request, context, nominee_link_id)
