+++ csss-site/src/elections/views/update_election/nominee_links/display_selected_election_nominee_links.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 
 from elections.models import NomineeLink
-from elections.views.create_context.nominee_links.update_election_nominee_links_html import \
+from elections.views.create_context.nominee_links.create_or_update_election.update_election_nominee_links_html import \
     create_context_for_update_election_nominee_links_html
 
 
@@ -22,4 +22,7 @@ def display_selected_election_and_nominee_links(request, election, context):
         election_date=election.date, election_time=election.date, election_type=election.election_type,
         websurvey_link=election.websurvey, create_new_election=election is None, election=election
     )
-    return render(request, 'elections/update_election/update_election_nominee_links.html', context)
+    return render(
+        request, 'elections/nominee_links/create_or_update_election/update_election_nominee_links.html',
+        context
+    )
