+++ csss-site/src/elections/views/update_election/nominee_links/display_selected_nominee__nominee_links.py
@@ -17,4 +17,7 @@ def display_current_nominee_link_election(request, context, nominee_link_id):
     render object that directs the user to the page for updating a nominee via nominee link
     """
     create_context_for_create_or_update_nominee__nominee_links_html(context, nominee_link_id=nominee_link_id)
-    return render(request, 'elections/update_nominee/create_or_update_nominee__nominee_links.html', context)
+    return render(
+        request, 'elections/nominee_links/create_or_update_nominee/create_or_update_nominee__nominee_links.html',
+        context
+    )
